refactor(domain): move GameInfo to it's own file, compile tests
